requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
blis==0.7.11
build==1.2.2
CacheControl==0.14.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
cleo==2.1.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
crashtest==0.4.1
crew==0.9.2
cryptography==43.0.0
cymem==2.0.8
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
durationpy==0.7
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google_search_results==2.4.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.5
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kubernetes==31.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langcodes==3.4.0
langserve==0.2.2
langsmith==0.1.80
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numba==0.60.0
numexpr==2.10.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.34.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pooch==1.8.2
posthog==3.6.6
preshed==3.0.9
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-pika==0.9.14
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.1
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
setuptools==70.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
sse-starlette==2.1.2
st-circular-progress==0.3.9
starlette==0.37.2
streamlit==1.35.0
sympy==1.13.3
tenacity==8.4.1
tesseract==0.1.3
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
trove-classifiers==2024.9.12
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.26.2
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.22.0
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.2
